refactor: modularizar funciones kanban
