+++ kanban/card.js
@@ -1,0 +1,145 @@
+// Funciones relacionadas con tarjetas y colores
+import { etapasImpresion } from '../firestore.js';
+import { openPedidoModal, completeStage } from '../pedidoModal.js';
+
+/**
+ * Genera un color HSL a partir de una cadena.
+ */
+export function stringToColor(str, s = 60, l = 80) {
+    let hash = 0;
+    for (let i = 0; i < str.length; i++) {
+        hash = str.charCodeAt(i) + ((hash << 5) - hash);
+    }
+    const h = Math.abs(hash) % 360;
+    return `hsl(${h}, ${s}%, ${l}%)`;
+}
+
+/**
+ * Devuelve un color pastel para cada etapa.
+ */
+export function etapaColumnColor(etapa) {
+    let hash = 0;
+    for (let i = 0; i < etapa.length; i++) {
+        hash = etapa.charCodeAt(i) + ((hash << 5) - hash);
+    }
+    const h = Math.abs(hash) % 360;
+    return `hsl(${h}, 70%, 93%)`;
+}
+
+/**
+ * Obtiene el color de fondo de una columna según el cliente predominante.
+ */
+export function getColumnColorByClientes(pedidosInEtapa) {
+    if (!pedidosInEtapa || pedidosInEtapa.length === 0) {
+        return 'hsl(210, 20%, 97%)';
+    }
+    const counts = {};
+    pedidosInEtapa.forEach(p => {
+        const cliente = p.cliente || '';
+        counts[cliente] = (counts[cliente] || 0) + 1;
+    });
+    let maxCliente = '';
+    let maxCount = 0;
+    Object.entries(counts).forEach(([cliente, count]) => {
+        if (count > maxCount) {
+            maxCount = count;
+            maxCliente = cliente;
+        }
+    });
+    const clienteColor = maxCount === 1
+        ? pedidosInEtapa[0].cliente || ''
+        : maxCliente;
+    if (!clienteColor) return 'hsl(210, 20%, 97%)';
+    return stringToColor(clienteColor, 90, 96);
+}
+
+/**
+ * Crea la tarjeta HTML para un pedido del Kanban.
+ */
+export function createKanbanCard(pedido) {
+    const card = document.createElement('div');
+    card.className = 'kanban-card';
+    card.id = `pedido-${pedido.id}`;
+    card.draggable = true;
+    card.dataset.id = pedido.id;
+    card.style.wordBreak = 'break-word';
+    card.style.overflowWrap = 'anywhere';
+    card.style.maxWidth = '100%';
+    card.style.minWidth = '0';
+
+    let fechaDisplay = '';
+    if (pedido.fecha) {
+        let fechaStr = pedido.fecha;
+        if (pedido.fecha instanceof Date) {
+            fechaStr = pedido.fecha.toISOString().slice(0, 10);
+        } else if (typeof pedido.fecha === 'string' && pedido.fecha.length >= 10) {
+            fechaStr = pedido.fecha.slice(0, 10);
+        }
+        fechaDisplay = ` (${fechaStr})`;
+    }
+
+    let clienteBadge = '';
+    if (pedido.cliente) {
+        const color = stringToColor(pedido.cliente);
+        clienteBadge = `<span class="badge badge-cliente ms-1" style="background:${color};color:#333;font-size:0.75em; margin-right:0.5em;">${pedido.cliente}</span>`;
+    }
+    const metrosBadge = pedido.metros
+        ? `<span class="badge bg-secondary ms-1" style="font-size:0.75em;">${pedido.metros} m</span>`
+        : '';
+
+    const etapasHtml = pedido.etapasSecuencia && pedido.etapasSecuencia.length > 0
+        ? `<div class="etapas-display">Secuencia: ${
+            pedido.etapasSecuencia
+                .filter(et => !etapasImpresion.includes(et))
+                .join(' -> ') || 'N/A'
+        }</div>`
+        : '';
+
+    let etapaBtnText = '';
+    let showEtapaBtn = false;
+    if (pedido.etapaActual !== 'Completado' && pedido.etapasSecuencia && pedido.etapasSecuencia.length > 0) {
+        const idx = pedido.etapasSecuencia.indexOf(pedido.etapaActual);
+        if (etapasImpresion.includes(pedido.etapaActual)) {
+            etapaBtnText = 'Iniciar trabajo';
+            showEtapaBtn = true;
+        } else if (idx > -1 && idx < pedido.etapasSecuencia.length - 1) {
+            etapaBtnText = 'Siguiente trabajo';
+            showEtapaBtn = true;
+        } else if (idx === pedido.etapasSecuencia.length - 1) {
+            etapaBtnText = 'Completar';
+            showEtapaBtn = true;
+        } else {
+            const lastEtapa = pedido.etapasSecuencia[pedido.etapasSecuencia.length - 1];
+            if (pedido.etapaActual !== lastEtapa) {
+                etapaBtnText = 'Siguiente trabajo';
+                showEtapaBtn = true;
+            }
+        }
+    }
+
+    card.innerHTML = `
+        <div class="kanban-card-header" style="word-break:break-word;overflow-wrap:anywhere;max-width:100%;">
+            ${clienteBadge}
+            <div>
+                ${metrosBadge}
+            </div>
+        </div>
+        <div class="kanban-card-body" style="word-break:break-word;overflow-wrap:anywhere;max-width:100%;">
+            <div style="font-weight:bold; font-size:1em; margin-bottom:0.2em;">
+                ${pedido.numeroPedido || 'N/A'}${fechaDisplay}
+            </div>
+            <p><strong>Máquina:</strong> ${pedido.maquinaImpresion || 'N/A'}</p>
+            ${etapasHtml}
+        </div>
+        <div class="kanban-card-footer" style="word-break:break-word;overflow-wrap:anywhere;max-width:100%;">
+            <button class="btn btn-sm btn-outline-primary" onclick="openPedidoModal('${pedido.id}')">Ver/Editar</button>
+            ${showEtapaBtn ? `<button class="btn btn-sm btn-outline-success mt-1" onclick="completeStage('${pedido.id}')">${etapaBtnText}</button>` : ''}
+        </div>
+    `;
+    return card;
+}
+
+// Exportación para pruebas en Node
+if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
+    module.exports = { getColumnColorByClientes };
+}
