+++ kanban/index.js
@@ -1,0 +1,4 @@
+// Archivo de entrada para el módulo Kanban
+export { renderKanban, renderJKanban } from './render.js';
+export { createKanbanCard, getColumnColorByClientes } from './card.js';
+export { setupKanbanScrolling, enableKanbanDragToScroll, trackBoardScroll, setContainerPosition, fixAllContainerTranslates } from './scroll.js';
