+++ package.json
@@ -4,10 +4,13 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-  "test": "jest"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs"
+  "type": "commonjs",
+  "dependencies": {
+    "jest": "^29.6.1"
+  }
 }
